# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Train API.",
        "title": "AI Train API",
        "contact": {},
        "license": {},
        "version": "0.2"
    },
    "host": "localhost:38080",
    "basePath": "/",
    "paths": {
...
        "/v1/proxy/changePW": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Change user password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "new password",
                        "name": "user_password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.PasswordInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.PlainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/introspection": {
            "post": {
                "description": "Get token meta information from provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Get token meta information from provider",
                "parameters": [
                    {
                        "description": "Introspection token",
                        "name": "Introspection_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.IntrospectionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.IntrospectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Revoke all tokens of a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Revoke all tokens of a user",
                "parameters": [
                    {
                        "description": "user token",
                        "name": "user_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.IntrospectionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.PlainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/query": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "query a existing user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "query a existing user information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.UserInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/refresh": {
            "post": {
                "description": "Refresh token with provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Refresh token with provider",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "refresh_token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.RefreshTokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.TokenResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/register": {
            "post": {
                "description": "Register a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "user information",
                        "name": "new_user_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.PlainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/token": {
            "post": {
                "description": "Exchange token from Provider",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Exchange token from Provider",
                "parameters": [
                    {
                        "description": "token request",
                        "name": "token_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.TokenReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.TokenResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/proxy/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a existing user information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Proxy"
                ],
                "summary": "Update a existing user information",
                "parameters": [
                    {
                        "description": "user information",
                        "name": "user_info",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.UpdatedUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.PlainResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "docs.GenericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "error response message"
                }
            }
        },
...
        "docs.IntrospectResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "client_id": {
                    "type": "string",
                    "example": "test_client_1"
                },
                "exp": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1530672296
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "scope": {
                    "type": "string",
                    "example": "read_write"
                },
                "token_type": {
                    "type": "string",
                    "example": "Bearer"
                },
                "username": {
                    "type": "string",
                    "example": "ogre0403@gmail.com"
                }
            }
        },
        "docs.IntrospectionReq": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "045e8bd5-58dc-4bd5-8254-dc3d1571c9cd"
                }
            }
        },
...
        "docs.PasswordInfo": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "654321"
                },
                "username": {
                    "type": "string",
                    "example": "jjj@abc.com"
                }
            }
        },
        "docs.PlainResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "response from proxy"
                }
            }
        },
        "docs.RefreshTokenReq": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "7e7f6442-09e0-44f3-a05b-d7ea516cc6c5"
                }
            }
        },
...
        "docs.TokenReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "f1bd4604-e5ce-469d-b3d7-5e1d8a3e87d7"
                }
            }
        },
        "docs.TokenResp": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "7e7f6442-09e0-44f3-a05b-d7ea516cc6c5"
                },
                "token": {
                    "type": "string",
                    "example": "045e8bd5-58dc-4bd5-8254-dc3d1571c9cd"
                }
            }
        },
...
        "docs.UpdatedUser": {
            "type": "object",
            "properties": {
                "cName": {
                    "type": "string",
                    "example": "莊莊莊"
                },
                "company": {
                    "type": "string",
                    "example": "國網中心"
                },
                "email-1": {
                    "type": "string",
                    "example": "ogre@ggg.com"
                },
                "email-2": {
                    "type": "string",
                    "example": "ogre@ggg.com"
                },
                "phone": {
                    "type": "string",
                    "example": "0911234567"
                },
                "text": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "jjj@abc.com"
                }
            }
        },
        "docs.UserInfo": {
            "type": "object",
            "properties": {
                "cName": {
                    "type": "string",
                    "example": "莊莊莊"
                },
                "company": {
                    "type": "string",
                    "example": "國網中心"
                },
                "email-1": {
                    "type": "string",
                    "example": "ogre@ggg.com"
                },
                "email-2": {
                    "type": "string",
                    "example": "ogre@ggg.com"
                },
                "password": {
                    "type": "string",
                    "example": "654321"
                },
                "phone": {
                    "type": "string",
                    "example": "0911234567"
                },
                "text": {
                    "type": "string"
                },
                "username": {
                    "type": "string",
                    "example": "jjj@abc.com"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}